+++ jdplus-main-base/jdplus-tramoseats-base-parent/jdplus-tramoseats-base-api/src/test/java/jdplus/tramoseats/base/api/tramoseats/TramoSeatsDictionariesTest.java
@@ -4,8 +4,6 @@
  */
 package jdplus.tramoseats.base.api.tramoseats;
 
-import jdplus.tramoseats.base.api.tramoseats.TramoSeatsDictionaries;
-
 /**
  *
  * @author PALATEJ